requirements.txt
nextcord~=2.0.0a3
python-dotenv~=0.19.0